# excourt_re/excourt_re-master/system/app/static/swagger.yaml
openapi: 3.0.0
info:
  title: ExCourt API
  description: API for court management system
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /upload/profile:
    post:
      summary: Upload a profile photo
      tags:
        - File Upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - photo
                - filename
              properties:
                photo:
                  type: string
                  format: binary
                  description: Profile photo file
                filename:
                  type: string
                  description: Name of the file
      responses:
        '201':
          description: Profile photo successfully uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  profileurl:
                    type: string
                    description: Path to the uploaded file
                  message:
                    type: string
        '500':
          description: Server error
  /upload/lostfind:
    post:
      summary: Upload a lost and found item photo
      tags:
        - File Upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - photo
              properties:
                photo:
                  type: string
                  format: binary
                  description: Lost and found item photo file
      responses:
        '201':
          description: Lost and found item photo successfully uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  profilename:
                    type: string
                    description: Generated filename for the uploaded photo
        '500':
          description: Server error
  /upload/find_profile:
    post:
      summary: Find a profile photo by student ID
      tags:
        - File Upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - student_id
              properties:
                student_id:
                  type: string
                  description: Student ID to search for
      responses:
        '200':
          description: Profile photo URL successfully found
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageUrl:
                    type: string
                    description: URL of the profile photo
        '400':
          description: Missing student ID
        '404':
          description: Profile photo not found
  /upload/find_code:
    post:
      summary: Find a QR code image by contact ID
      tags:
        - File Upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - con_id
              properties:
                con_id:
                  type: string
                  description: Contact ID to search for
      responses:
        '200':
          description: QR code image URL successfully found
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageUrl:
                    type: string
                    description: URL of the QR code image
        '400':
          description: Missing contact ID
        '404':
          description: QR code image not found
  /upload/find_lost:
    post:
      summary: Find a lost item photo by lost item ID
      tags:
        - File Upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Lost_id
              properties:
                Lost_id:
                  type: integer
                  description: Lost item ID to search for
      responses:
        '200':
          description: Lost item photo URL successfully found
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageUrl:
                    type: string
                    description: URL of the lost item photo
        '400':
          description: Missing lost item ID
        '404':
          description: Lost item photo not found
        '500':
          description: Server error
  /upload/profiles/{filename}:
    get:
      summary: Get a profile photo by filename
      tags:
        - File Upload
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
          description: Name of the profile photo file
      responses:
        '200':
          description: Profile photo successfully retrieved
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Profile photo not found
  /upload/code/{filename}:
    get:
      summary: Get a QR code image by filename
      tags:
        - File Upload
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
          description: Name of the QR code image file
      responses:
        '200':
          description: QR code image successfully retrieved
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: QR code image not found
  /upload/lost/{filename}:
    get:
      summary: Get a lost item photo by filename
      tags:
        - File Upload
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
          description: Name of the lost item photo file
      responses:
        '200':
          description: Lost item photo successfully retrieved
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Lost item photo not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [] 